Update data_analysis.rs
Created a public module (data_analysis) which will hold everything in the file for future use. Also created a function to calculate basic statistics for the dataset such as the averages of certain statistics and the number of houses in the set. It returns a hashmap of the average prices, number of bedrooms, and number of bathrooms.